+++ src/main/java/org/example/blood_help_app/repository/interfaces/IBloodUnitRepository.java
@@ -2,5 +2,5 @@ package org.example.blood_help_app.repository.interfaces;
 
 import org.example.blood_help_app.domain.donationsdata.BloodUnit;
 
-public interface IBloodUnitRepository extends IRepository<Long, BloodUnit> {
+public interface IBloodUnitRepository extends IRepository<Integer, BloodUnit> {
 }
