+++ src/main/java/org/example/blood_help_app/repository/interfaces/IDoctorRepository.java
@@ -2,5 +2,5 @@ package org.example.blood_help_app.repository.interfaces;
 
 import org.example.blood_help_app.domain.users.Doctor;
 
-public interface IDoctorRepository extends IRepository<Long, Doctor> {
+public interface IDoctorRepository extends IRepository<Integer, Doctor> {
 }
