+++ src/main/java/org/example/blood_help_app/repository/interfaces/IDonorRepository.java
@@ -2,5 +2,5 @@ package org.example.blood_help_app.repository.interfaces;
 
 import org.example.blood_help_app.domain.users.Donor;
 
-public interface IDonorRepository extends IRepository<Long, Donor> {
+public interface IDonorRepository extends IRepository<Integer, Donor> {
 }
