+++ src/main/java/org/example/blood_help_app/repository/interfaces/IUserRepository.java
@@ -4,6 +4,6 @@ import org.example.blood_help_app.domain.users.User;
 
 import java.util.Optional;
 
-public interface IUserRepository extends IRepository<Long, User> {
+public interface IUserRepository extends IRepository<Integer, User> {
     Optional<User> findByCredentials(String email, String encryptedPassword);
 }
